fraudfinder/ops/server/requirements.txt
google-cloud-aiplatform
google-cloud-storage
gunicorn
Flask
